python/requirements.txt
adafruit-blinka==8.12.0
adafruit-circuitpython-bno055
bottle==0.12.25
cheroot==10.0.0
dataclasses_json==0.6.7
gpsdclient==1.3.2
grpcio==1.64.1
json5==0.9.25
luma.oled==3.12.0
luma.lcd==2.11.0
pillow==10.4.0
numpy==1.26.2
pandas==1.5.3
pydeepskylog==1.3.2
pyjwt==2.8.0
python-libinput==0.3.0a0
pytz==2022.7.1
requests==2.28.2
rpi-hardware-pwm==0.1.4
scipy
scikit-learn==1.2.2
sh==1.14.3
skyfield==1.45
timezonefinder==6.1.9
tqdm==4.65.0
protobuf==4.25.2
aiofiles==24.1.0
